requirements.txt
numpy>=1.14.3
scipy>=1.1.0
pandas>=0.23.0
xlrd>=1.1.0
openpyxl>=2.5.0
scikit-learn>=0.18.1
future>=0.16.0
Keras>=2.2.2
tensorflow>=1.10.1
PyYAML>=3.13